third_party/WebKit/LayoutTests/media/controls/controls-video-keynav.html

<DOCTYPE html>
<title>Test media controls video keyboard navigation</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../media-file.js"></script>
<script src="../media-controls.js"></script>
<video controls></video>
<script>
test(_ => {
  assert_true('eventSender' in window);
}, 'eventSender required');

async_test(function(t) {
    var video = document.querySelector("video");
    video.src = findMediaFile("video", "../content/test");
    assert_equals(video.volume, 1);
    assert_equals(video.currentTime, 0);

    // Focus the video.
    video.focus();

    video.oncanplaythrough = t.step_func_done(function() {
      video.onplay = t.step_func(() => {
        assert_false(video.paused);
        video.onpause = t.step_func(() => {
          assert_true(video.paused);
          var cur = video.currentTime;
          video.onseeked = t.step_func(function() {
            assert_greater_than(video.currentTime, cur);
            video.onseeked = t.step_func(function() {
              assert_equals(video.currentTime, video.duration);
              video.onseeked = t.step_func(function() {
                assert_less_than(video.currentTime, video.duration);
                video.onseeked = t.step_func(function() {
                  assert_equals(video.currentTime, 0);
                });
                // 'Home' sets the timeline to 0.
                eventSender.keyDown("Home");
              });
              // 'Left' scrubs the timeline back.
              eventSender.keyDown("ArrowLeft");
            });
            // 'End' sets the timeline to end.
            eventSender.keyDown("End");
          });
          // 'Right' scrubs the timeline forward.
          eventSender.keyDown("ArrowRight");
        });
        // 'Enter' again pauses the video.
        eventSender.keyDown("Enter");
      });
      // 'Enter' plays the video.
      eventSender.keyDown("Enter");
    });
});
</script>
